tests for liquidation

+++ src/types.rs
@@ -96,7 +96,7 @@ pub struct Order {
     pub withdraw_collateral_amount: TokenAmount,
 
     /// Target leverage X for this step, e.g. 5 means 5x.
-    pub target_leverage_x: i64,
+    pub target_leverage_x: u32,
 
     pub created_at: Timestamp,
     pub valid_from: Timestamp,
